refactor jsapi - http
