+++ src/logbypass/http.h
@@ -1,7 +1,6 @@
 #ifndef XPROFILER_SRC_LOGBYPASS_HTTP_H
 #define XPROFILER_SRC_LOGBYPASS_HTTP_H
 
-#include "nan.h"
 #include "xpf_mutex-inl.h"
 
 namespace xprofiler {
@@ -24,13 +23,6 @@ struct HttpStatistics {
 
 void WriteHttpStatus(EnvironmentData* env_data, bool log_format_alinode,
                      uint32_t http_patch_timeout);
-
-// javascript-accessible
-void AddLiveRequest(const Nan::FunctionCallbackInfo<v8::Value>& info);
-void AddCloseRequest(const Nan::FunctionCallbackInfo<v8::Value>& info);
-void AddSentRequest(const Nan::FunctionCallbackInfo<v8::Value>& info);
-void AddRequestTimeout(const Nan::FunctionCallbackInfo<v8::Value>& info);
-void AddHttpStatusCode(const Nan::FunctionCallbackInfo<v8::Value>& info);
 }  // namespace xprofiler
 
 #endif /* XPROFILER_SRC_LOGBYPASS_HTTP_H */
